+++ src/index.js
@@ -1,8 +1,8 @@
-import EntityFactory from "./EntityFactory"
-import InorigoEnums from "./Enums"
-import InorigoAPI from "./InorigoApi"
-import Parser from "./Parser"
-import Entity from "./Entity"
-import EntityRelation from "./EntityRelation"
+import { InorigoAPI } from "./API/InorigoApi"
+import { Entity } from "./Entity/Entity"
+import { EntityRelation } from "./Entity/EntityRelation"
+import { EntityFactory } from "./Entity/EntityFactory"
+import { InorigoEnums } from "./Utils/Enums"
+import { Parser } from "./Utils/Parser"
 
 export { EntityFactory, InorigoEnums, InorigoAPI, Parser, Entity, EntityRelation }
